+++ index.ios.js
@@ -3,7 +3,6 @@ import "./src/lib/ErrorReporting"
 
 import Containers from "./src/lib/Containers"
 import Components from "./src/lib/Components"
-import Routes from "./src/lib/relay/routes"
 
 import "./src/lib/relay/config"
 import "./src/lib/AppRegistry"
@@ -11,5 +10,4 @@ import "./src/lib/AppRegistry"
 export default {
   Containers,
   Components,
-  Routes,
 }