+++ src/lib/Components/ArtworkGrids/RelayConnections/ArtistForSaleArtworksGrid.tsx
@@ -6,27 +6,23 @@ import Artwork from "../Artwork"
 // This is so that TypeScript won’t remove the seemingly unused `Artwork` import. Relay depends on it to exist.
 Artwork
 
-const GeneArtworksGrid = createPaginationContainer(
+const ArtistForSaleArtworksGrid = createPaginationContainer(
   InfiniteScrollArtworksGrid,
   {
-    gene: graphql.experimental`
-      fragment GeneArtworksGrid_gene on Gene
+    artist: graphql.experimental`
+      fragment ArtistForSaleArtworksGrid_artist on Artist
         @argumentDefinitions(
           count: { type: "Int", defaultValue: 10 }
           cursor: { type: "String" }
-          sort: { type: "String", defaultValue: "-partner_updated_at" }
-          medium: { type: "String", defaultValue: "*" }
-          priceRange: { type: "String", defaultValue: "*-*" }
+          filter: { type: "[ArtistArtworksFilters]", defaultValue: [IS_FOR_SALE] }
         ) {
         __id
-        artworks: artworks_connection(
+        forSaleArtworks: artworks_connection(
           first: $count
           after: $cursor
-          sort: $sort
-          medium: $medium
-          price_range: $priceRange
-          for_sale: true
-        ) @connection(key: "GeneArtworksGrid_artworks") {
+          filter: $filter
+          sort: partner_updated_at_desc
+        ) @connection(key: "ArtistForSaleArtworksGrid_forSaleArtworks") {
           pageInfo {
             hasNextPage
             startCursor
@@ -48,7 +44,7 @@ const GeneArtworksGrid = createPaginationContainer(
   {
     direction: "forward",
     getConnectionFromProps(props) {
-      return props.gene && props.gene.artworks
+      return props.artist && props.artist.forSaleArtworks
     },
     getFragmentVariables(prevVars, totalCount) {
       return {
@@ -56,27 +52,24 @@ const GeneArtworksGrid = createPaginationContainer(
         count: totalCount,
       }
     },
-    getVariables(props, { count, cursor }, fragmentVariables) {
+    getVariables(props, { count, cursor }, { filter }) {
       return {
-        ...fragmentVariables,
-        __id: props.gene.__id,
+        __id: props.artist.__id,
         count,
         cursor,
+        filter,
       }
     },
     query: graphql.experimental`
-      query GeneArtworksGridQuery(
+      query ArtistForSaleArtworksGridQuery(
         $__id: ID!
         $count: Int!
         $cursor: String
-        $sort: String
-        $medium: String
-        $priceRange: String
+        $filter: [ArtistArtworksFilters]
       ) {
         node(__id: $__id) {
-          ... on Gene {
-            ...GeneArtworksGrid_gene
-              @arguments(count: $count, cursor: $cursor, sort: $sort, medium: $medium, priceRange: $priceRange)
+          ... on Artist {
+            ...ArtistForSaleArtworksGrid_artist @arguments(count: $count, cursor: $cursor, filter: $filter)
           }
         }
       }
@@ -84,9 +77,9 @@ const GeneArtworksGrid = createPaginationContainer(
   }
 )
 
-export default GeneArtworksGrid
+export default ArtistForSaleArtworksGrid
 
-export interface GeneRelayProps {
+export interface ArtistRelayProps {
   artist: {
     artworks_connection: {
       pageInfo: {