+++ src/lib/Components/Consignments/Components/BottomAlignedButton.tsx
@@ -25,15 +25,14 @@ export interface ButtonBodyStyle {
   paddingTop: number
 }
 
-export interface BottomAlignedProps {
+export interface BottomAlignedProps extends React.Props<JSX.Element> {
   onPress: () => void
-  children?: any[]
   bodyStyle: ButtonBodyStyle
   buttonText: string
   disabled?: boolean
 }
 
-const render = (props: BottomAlignedProps) =>
+const BottomAlignedButton: React.SFC<BottomAlignedProps> = props =>
   <KeyboardAvoidingView behavior="padding" keyboardVerticalOffset={60} style={{ flex: 1 }}>
     <View key="space-eater" style={{ flexGrow: 1 }}>
       {props.children}
@@ -47,8 +46,4 @@ const render = (props: BottomAlignedProps) =>
     </TouchableOpacity>
   </KeyboardAvoidingView>
 
-export default class BottomAlignedButton extends React.Component<BottomAlignedProps, null> {
-  render() {
-    return render(this.props)
-  }
-}
+export default BottomAlignedButton