+++ src/lib/Components/Inbox/Conversations/__tests__/ZeroStateInbox-tests.tsx
@@ -3,14 +3,18 @@ import * as React from "react"
 import "react-native"
 import * as renderer from "react-test-renderer"
 
-import Inbox from "../"
+import { Conversations } from "../"
 
 it("looks correct when the user has no conversations", () => {
-  const tree = renderer.create(<Inbox me={{ conversations: { edges: [] } }} />).toJSON()
+  const tree = renderer
+    .create(<Conversations me={{ conversations: { pageInfo: { hasNextPage: false }, edges: [] } }} />)
+    .toJSON()
   expect(tree).toMatchSnapshot()
 })
 
 it("renders correct when the user has no conversations", () => {
-  const tree = renderer.create(<Inbox me={{ conversations: { edges: [] } }} />).toJSON()
+  const tree = renderer
+    .create(<Conversations me={{ conversations: { pageInfo: { hasNextPage: false }, edges: [] } }} />)
+    .toJSON()
   expect(tree).toMatchSVGSnapshot(480, 1024)
 })