+++ src/lib/Components/Text/Headline.tsx
@@ -4,8 +4,9 @@ import { StyleSheet, Text, TextProperties } from "react-native"
 export default class Headline extends React.Component<TextProperties, any> {
   render() {
     const content = (this.props.children || "") as string
+
     return (
-      <Text {...this.props} style={[styles.default, this.props.style, styles.required]}>
+      <Text style={[styles.default, this.props.style, styles.required]}>
         {content.toUpperCase()}
       </Text>
     )