+++ src/lib/Containers/__stories__/Artist.story.tsx
@@ -1,24 +1,47 @@
 import { storiesOf } from "@storybook/react-native"
 import * as React from "react"
-import { RootContainer } from "react-relay"
 
-import Routes from "../../relay/routes"
 import Artist from "../Artist"
+
+// TODO Move to metametaphysics after Relay Modern migration
+import { graphql, QueryRenderer } from "react-relay"
+import createEnvironment from "../../relay/createEnvironment"
+const RootContainer: React.SFC<any> = ({ Component, artistID }) => {
+  return (
+    <QueryRenderer
+      environment={createEnvironment()}
+      query={graphql`
+        query ArtistQuery($artistID: String!) {
+          artist(id: $artistID) {
+            ...Artist_artist
+          }
+        }
+      `}
+      variables={{
+        artistID,
+      }}
+      render={({ error, props }) => {
+        if (error) {
+          console.error(error)
+        } else if (props) {
+          return <Component {...props} />
+        }
+        return null
+      }}
+    />
+  )
+}
 
 storiesOf("Artist/Relay")
   .add("Glenn Brown", () => {
-    const artistRoute = new Routes.Artist({ artistID: "glenn-brown" })
-    return <RootContainer Component={Artist} route={artistRoute} />
+    return <RootContainer Component={Artist} artistID="glenn-brown" />
   })
   .add("Leda Catunda", () => {
-    const artistRoute = new Routes.Artist({ artistID: "leda-catunda" })
-    return <RootContainer Component={Artist} route={artistRoute} />
+    return <RootContainer Component={Artist} artistID="glenn-brown" />
   })
   .add("Jorge Vigil", () => {
-    const artistRoute = new Routes.Artist({ artistID: "jorge-vigil" })
-    return <RootContainer Component={Artist} route={artistRoute} />
+    return <RootContainer Component={Artist} artistID="glenn-brown" />
   })
   .add("Rita Maas", () => {
-    const artistRoute = new Routes.Artist({ artistID: "rita-maas" })
-    return <RootContainer Component={Artist} route={artistRoute} />
+    return <RootContainer Component={Artist} artistID="glenn-brown" />
   })