+++ src/lib/relay/config.tsx
@@ -1,5 +1,4 @@
 import { NativeModules } from "react-native"
-import * as Relay from "react-relay"
 const { Emission } = NativeModules
 
 let metaphysicsURL
@@ -15,22 +14,12 @@ if (Emission && Emission.gravityAPIHost && Emission.metaphysicsAPIHost) {
 
 export { metaphysicsURL, gravityURL }
 
-if (Emission) {
-  Relay.injectNetworkLayer(
-    new Relay.DefaultNetworkLayer(metaphysicsURL, {
-      headers: {
-        "X-USER-ID": Emission.userID,
-        "X-ACCESS-TOKEN": Emission.authenticationToken,
-      },
-    })
-  )
-}
-
 // Disable the native polyfill during development, which will make network requests show-up in the Chrome dev-tools.
 // Specifically, in our case, we get to see the Relay requests.
 //
 // It will be `undefined` unless running inside Chrome.
 //
+declare var global: any
 if (__DEV__ && global.originalXMLHttpRequest !== undefined) {
   global.XMLHttpRequest = global.originalXMLHttpRequest
   // tslint:disable-next-line:no-var-requires